fortran typecast to (INT) since f interface is strictly int for now

+++ fortran/finufft_f.cpp
@@ -11,35 +11,35 @@ void finufft1d1_f_(int *nj,double* xj,dcomplex* cj,int *iflag, double *eps,
 		  int *ms, dcomplex* fk, int *ier)
 {
   nufft_opts opts;
-  *ier = finufft1d1(*nj,xj,cj,*iflag,*eps,*ms,fk,opts);
+  *ier = finufft1d1((INT)*nj,xj,cj,*iflag,*eps,(INT)*ms,fk,opts);
 }
 
 void finufft1d2_f_(int *nj,double* xj,dcomplex* cj,int *iflag, double *eps,
 		  int *ms, dcomplex* fk, int *ier)
 {
   nufft_opts opts;
-  *ier = finufft1d2(*nj,xj,cj,*iflag,*eps,*ms,fk,opts);
+  *ier = finufft1d2((INT)*nj,xj,cj,*iflag,*eps,(INT)*ms,fk,opts);
 }
 
 void finufft1d3_f_(int *nj,double* xj,dcomplex* cj,int *iflag, double *eps,
 		  int *nk, double* s, dcomplex* fk, int *ier)
 {
   nufft_opts opts;
-  *ier = finufft1d3(*nj,xj,cj,*iflag,*eps,*nk,s,fk,opts);
+  *ier = finufft1d3((INT)*nj,xj,cj,*iflag,*eps,(INT)*nk,s,fk,opts);
 }
 
 void finufft2d1_f_(int *nj,double* xj,double *yj,dcomplex* cj,int *iflag,
 		   double *eps, int *ms, int *mt, dcomplex* fk, int *ier)
 {
   nufft_opts opts;
-  *ier = finufft2d1(*nj,xj,yj,cj,*iflag,*eps,*ms,*mt,fk,opts);
+  *ier = finufft2d1((INT)*nj,xj,yj,cj,*iflag,*eps,(INT)*ms,(INT)*mt,fk,opts);
 }
 
 void finufft2d2_f_(int *nj,double* xj,double *yj,dcomplex* cj,int *iflag,
 		   double *eps, int *ms, int *mt, dcomplex* fk, int *ier)
 {
   nufft_opts opts;
-  *ier = finufft2d2(*nj,xj,yj,cj,*iflag,*eps,*ms,*mt,fk,opts);
+  *ier = finufft2d2((INT)*nj,xj,yj,cj,*iflag,*eps,(INT)*ms,(INT)*mt,fk,opts);
 }
 
 void finufft2d3_f_(int *nj,double* xj,double* yj, dcomplex* cj,int *iflag,
@@ -47,7 +47,7 @@ void finufft2d3_f_(int *nj,double* xj,double* yj, dcomplex* cj,int *iflag,
 		   int *ier)
 {
   nufft_opts opts;
-  *ier = finufft2d3(*nj,xj,yj,cj,*iflag,*eps,*nk,s,t,fk,opts);
+  *ier = finufft2d3((INT)*nj,xj,yj,cj,*iflag,*eps,(INT)*nk,s,t,fk,opts);
 }
 
 void finufft3d1_f_(int *nj,double* xj,double *yj,double* zj,dcomplex* cj,
@@ -55,7 +55,7 @@ void finufft3d1_f_(int *nj,double* xj,double *yj,double* zj,dcomplex* cj,
 		   dcomplex* fk, int *ier)
 {
   nufft_opts opts;
-  *ier = finufft3d1(*nj,xj,yj,zj,cj,*iflag,*eps,*ms,*mt,*mu,fk,opts);
+  *ier = finufft3d1((INT)*nj,xj,yj,zj,cj,*iflag,*eps,(INT)*ms,(INT)*mt,(INT)*mu,fk,opts);
 }
 
 void finufft3d2_f_(int *nj,double* xj,double *yj,double* zj,dcomplex* cj,
@@ -63,7 +63,7 @@ void finufft3d2_f_(int *nj,double* xj,double *yj,double* zj,dcomplex* cj,
 		   dcomplex* fk, int *ier)
 {
   nufft_opts opts;
-  *ier = finufft3d2(*nj,xj,yj,zj,cj,*iflag,*eps,*ms,*mt,*mu,fk,opts);
+  *ier = finufft3d2((INT)*nj,xj,yj,zj,cj,*iflag,*eps,(INT)*ms,(INT)*mt,(INT)*mu,fk,opts);
 }
 
 void finufft3d3_f_(int *nj,double* xj,double* yj, double*zj, dcomplex* cj,
@@ -71,5 +71,5 @@ void finufft3d3_f_(int *nj,double* xj,double* yj, double*zj, dcomplex* cj,
 		   double* u, dcomplex* fk, int *ier)
 {
   nufft_opts opts;
-  *ier = finufft3d3(*nj,xj,yj,zj,cj,*iflag,*eps,*nk,s,t,u,fk,opts);
+  *ier = finufft3d3((INT)*nj,xj,yj,zj,cj,*iflag,*eps,(INT)*nk,s,t,u,fk,opts);
 }
